I18n works in an import file, so reviews schema i18n was moved to /imports/startup/client/i18n.js.
